Add validation with react-hook-form and yup
As react-hook-form handles both changes in the inputs and
validation of the data, it was easy to build the form and
implemnet validation right away. Yup is used to create a validation
schema.

Submitting the form logs an object to the console that is formatted
as the API expects it.